Initial commit of potential a11y changes to core.
Note that this branch will never be directly merged into mainline, and
is only meant to be used as a very basic means of sharing context and
functionality of ongoing a11y work.

Note that these changes are tied to other changes in the Blockly
keyboard experimentation plugin, and both changes are needed together in
order to have a working a11y with screenreader experience.

This commit also includes many changes from attempts to bring a11y
support to specific areas (such as input fields) and to make core
tab-navigable on its own (though this was abandoned in favor of relying
on keyboard navigation, instead).